requirements.txt
fastapi==0.115.12
numpy==2.2.6
pydantic==2.11.5
PyPDF2==3.0.1
python-dotenv==1.1.0
Requests==2.32.3
sentence_transformers==4.1.0
streamlit==1.45.1
uvicorn[standard]==0.24.0
ctransformers